Added an OutletInitializer as a helper initializer class
