+++ xlb/helper/__init__.py
@@ -1,5 +1,6 @@
 from xlb.helper.nse_fields import create_nse_fields
 from xlb.helper.initializers import initialize_eq
 from xlb.helper.initializers import initialize_multires_eq
+from xlb.helper.initializers import OutletInitializer
 from xlb.helper.check_boundary_overlaps import check_bc_overlaps
 from xlb.helper.simulation_manager import MultiresSimulationManager
